+++ test/test_grade/test_integration.py
@@ -14,6 +14,7 @@ from unittest import mock
 
 from otter.generate import main as generate
 from otter.grade import main as grade
+from otter.grade.utils import POINTS_POSSIBLE_LABEL
 from otter.run.run_autograder.autograder_config import AutograderConfig
 from otter.utils import loggers
 
@@ -168,6 +169,9 @@ def test_notebooks_with_pdfs(expected_points):
     )
     assert sorted(dir1_contents) == sorted(dir2_contents), f"'{FILE_MANAGER.get_path('notebooks/')}' and 'test/submission_pdfs' have different contents"
 
+    # check that the row with point totals for each question exists
+    assert any(POINTS_POSSIBLE_LABEL in row for row in df_test.itertuples(index=False))
+
 
 @mock.patch("otter.grade.launch_containers")
 def test_single_notebook_grade(mocked_launch_grade):
@@ -182,7 +186,19 @@ def test_single_notebook_grade(mocked_launch_grade):
         "q6": 5.0,
         "q2b": 2.0,
         "q7": 1.0,
-        "percent_correct": 1.0,
+        "percent_correct": float('nan'),
+        "total_points_earned": 15.0,
+        "file": POINTS_POSSIBLE_LABEL,
+    },{
+        "q1": 2.0,
+        "q2": 2.0,
+        "q3": 2.0,
+        "q4": 1.0,
+        "q6": 4.0,
+        "q2b": 2.0,
+        "q7": 1.0,
+        "percent_correct": 0.933333,
+        "total_points_earned": 14.0,
         "file": "passesAll.ipynb",
     }])
 
@@ -211,7 +227,7 @@ def test_single_notebook_grade(mocked_launch_grade):
     )
 
     mocked_launch_grade.assert_called_with(notebook_path, [notebook_path], **kw_expected)
-    assert output == 1.0
+    assert output == 0.933333
 
 
 @mock.patch("otter.grade.launch_containers")
@@ -227,7 +243,19 @@ def test_config_overrides(mocked_launch_grade):
         "q6": 5.0,
         "q2b": 2.0,
         "q7": 1.0,
+        "percent_correct": float('nan'),
+        "total_points_earned": 15.0,
+        "file": POINTS_POSSIBLE_LABEL,
+    },{
+        "q1": 2.0,
+        "q2": 2.0,
+        "q3": 2.0,
+        "q4": 1.0,
+        "q6": 5.0,
+        "q2b": 2.0,
+        "q7": 1.0,
         "percent_correct": 1.0,
+        "total_points_earned": 15.0,
         "file": "passesAll.ipynb",
     }])]
 
@@ -281,12 +309,23 @@ def test_config_overrides_integration():
         "q6": 5.0,
         "q2b": 2.0,
         "q7": 1.0,
+        "percent_correct": float('nan'),
+        "total_points_earned": 13.0,
+        "file": POINTS_POSSIBLE_LABEL,
+    },{
+        "q1": 0.0,
+        "q2": 2.0,
+        "q3": 2.0,
+        "q4": 1.0,
+        "q6": 5.0,
+        "q2b": 2.0,
+        "q7": 1.0,
         "percent_correct": 1.0,
+        "total_points_earned": 13.0,
         "file": os.path.splitext(os.path.basename(ZIP_SUBM_PATH))[0],
     }])
 
     # Sort the columns by label so the dataframes can be compared with ==.
     got = got.reindex(sorted(got.columns), axis=1)
     want = want.reindex(sorted(want.columns), axis=1)
-    
     assert got.equals(want)